% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpAggRates.R
\name{aggAR}
\alias{aggAR}
\title{Aggregates attack-rates from time step to annual}
\usage{
aggAR(A, inStep, outStep)
}
\arguments{
\item{A}{A vector of attack-rates}

\item{inStep}{Input interval size in days}

\item{outStep}{output interval size in days}
}
\description{
Aggregates attack-rates from time step to annual
}
