% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2pr.R
\name{AR2PR}
\alias{AR2PR}
\title{Calculate the prevalence rate and treatment rate for a given attack-rate,
model, and set of model parameters}
\usage{
AR2PR(A, PAR, Bfn)
}
\arguments{
\item{A}{An attack-rate}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Calculate the prevalence rate and treatment rate for a given attack-rate,
model, and set of model parameters
}
