% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_sim.R
\name{AR2PRopt}
\alias{AR2PRopt}
\title{Helper function for finding the maximum prevelance for a given model and set
of paramaters}
\usage{
AR2PRopt(A, PAR, Bfn)
}
\arguments{
\item{A}{An attack-rate}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Helper function for finding the maximum prevelance for a given model and set
of paramaters
}
