% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_sim.R
\name{findAXmax}
\alias{findAXmax}
\title{Finds the maximum prevalence and associated attack-rate for a given model and
set of parameters}
\usage{
findAXmax(PAR, Bfn)
}
\arguments{
\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Finds the maximum prevalence and associated attack-rate for a given model and
set of parameters
}
