% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2pr.R
\name{findYeq}
\alias{findYeq}
\title{Calculates the equilibrium Y for a given attack-rate and model}
\usage{
findYeq(PAR, Bfn)
}
\arguments{
\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Calculates the equilibrium Y for a given attack-rate and model
}
