% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sim.R
\name{fn}
\alias{fn}
\title{Helper objective function for finding the optimal attack-rate for a given
prevalence rate}
\usage{
fn(A, PAR, Bfn, X)
}
\arguments{
\item{A}{An attack-rate}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}

\item{X}{A prevalence rate}
}
\description{
Helper objective function for finding the optimal attack-rate for a given
prevalence rate
}
