% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sim.R
\name{fn2}
\alias{fn2}
\title{Helper objective function for finding a pair of optimal attack-rates for a
given pair of starting prevalence rates when the model has an exposure class}
\usage{
fn2(params, PAR, Bfn, X)
}
\arguments{
\item{params}{A pair of attack-rates}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}

\item{X}{A pair of prevalence rates}
}
\description{
Helper objective function for finding a pair of optimal attack-rates for a
given pair of starting prevalence rates when the model has an exposure class
}
