% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpAggRates.R
\name{interpRates}
\alias{interpRates}
\title{Interpolate a vector of rates to match the model time step}
\usage{
interpRates(rates, inStep = 365, outStep = 10)
}
\arguments{
\item{rates}{A vector of rates}

\item{inStep}{Input interval size in days}

\item{outStep}{output interval size in days}
}
\description{
Interpolate a vector of rates to match the model time step
}
