% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeB}
\alias{makeB}
\title{A function for making the matrix representation of the simple SIS model}
\usage{
makeB(PAR)
}
\arguments{
\item{PAR}{A set of model parameters}
}
\description{
A function for making the matrix representation of the simple SIS model
}
