% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeBage}
\alias{makeBage}
\title{A function for making the matrix representation of the age of infection (AoI)
with overwriting superinfection}
\usage{
makeBage(PAR)
}
\arguments{
\item{PAR}{A set of model parameters}
}
\description{
A function for making the matrix representation of the age of infection (AoI)
with overwriting superinfection
}
