% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeBdrugs}
\alias{makeBdrugs}
\title{A function for making the matrix representation of the simple  model with
drugs added}
\usage{
makeBdrugs(PAR)
}
\arguments{
\item{PAR}{A set of model parameters}
}
\description{
A function for making the matrix representation of the simple  model with
drugs added
}
