% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeD}
\alias{makeD}
\title{Generic function for making the D matrix that maps the state vector, Y, to
prevalence}
\usage{
makeD(PAR)
}
\arguments{
\item{PAR}{A set of model parameters}
}
\description{
Generic function for making the D matrix that maps the state vector, Y, to
prevalence
}
