% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeW}
\alias{makeW}
\title{Generic function for making the W matrix associated an input B function and
set of parameters}
\usage{
makeW(PAR, Bfn)
}
\arguments{
\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Generic function for making the W matrix associated an input B function and
set of parameters
}
