% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sim.R
\name{optimA}
\alias{optimA}
\title{A function for finding the optimal attack-rate for a given prevalence rate}
\usage{
optimA(X, PAR, Bfn)
}
\arguments{
\item{X}{A prevalence rate}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
A function for finding the optimal attack-rate for a given prevalence rate
}
