% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAR2PR.R
\name{plotAR2PR}
\alias{plotAR2PR}
\title{Make plots showing the relationship between AR and PR for a number of different Tx levels}
\usage{
plotAR2PR(PAR, Bfn, AR_range = c(0.01, 1), logX = T, logY = F,
  plotC = T, plotXover1_C = F)
}
\arguments{
\item{PAR}{Model parameters}

\item{Bfn}{Function for the model}

\item{AR_range}{Min and max of the range of AR plotted}

\item{logX}{Toggle for taking the log of the x-axis}

\item{logY}{Toggle for taking the log of the y-axis}

\item{plotC}{Toggle for plotting the proportion treated}

\item{plotXover1_C}{Toggle for plotting PR relative to those not on Tx}
}
\description{
Make plots showing the relationship between AR and PR for a number of different Tx levels
}
