% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2ar.R
\name{PR2AR}
\alias{PR2AR}
\title{Master function for converting prevalence rates into attack-rates}
\usage{
PR2AR(X, PAR, eq = F, showMessages = F, cpp = T)
}
\arguments{
\item{X}{A vector of prevalence rates}

\item{PAR}{A set of parameters for the mechanistic model of malaria}

\item{eq}{Toggle for whether to calculate all attack-rates at equilibrium or
to only calculate the initial states at equilibrium and simulate forward}

\item{showMessages}{Toggle for whether the function prints messages}

\item{cpp}{A toggle for using the C++ version of the forward simulation
algorithm}
}
\description{
Master function for converting prevalence rates into attack-rates
}
