% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_sim.R
\name{PR2AReq}
\alias{PR2AReq}
\title{Finds the equilibrium attack-rates for an input vector of prevalences,
conditional on the chosen model and parameters}
\usage{
PR2AReq(X, PAR, Bfn, showMessages = F)
}
\arguments{
\item{X}{A vector of prevalence rates}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}

\item{showMessages}{Controls whether or not the function prints messages}
}
\description{
Finds the equilibrium attack-rates for an input vector of prevalences,
conditional on the chosen model and parameters
}
