% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sim.R
\name{simAR}
\alias{simAR}
\title{Generic forward simulation function that can accomdate models with and
without exposure classes}
\usage{
simAR(X, PAR, Bfn)
}
\arguments{
\item{X}{A vector of prevalence rates}

\item{PAR}{A set of model parameters}

\item{Bfn}{A function for building the matrix represention of the system of
equations that update the state vector}
}
\description{
Generic forward simulation function that can accomdate models with and
without exposure classes
}
