% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggAR.R
\name{aggAR}
\alias{aggAR}
\title{Aggregates attack-rates from time step to annual}
\usage{
aggAR(A, dt)
}
\arguments{
\item{A}{A vector of attack-rates}

\item{dt}{length of time step}
}
\description{
Aggregates attack-rates from time step to annual
}
